requirements.txt
aioboto3==13.1.1
aiobotocore==2.13.1
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioitertools==0.12.0
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.0
asyncio==3.4.3
attrs==24.2.0
billiard==4.2.1
boto3==1.34.131
botocore==1.34.131
celery==5.4.0
certifi==2024.8.30
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
fastapi==0.115.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.4.2
MarkupSafe==3.0.1
multidict==6.1.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pluggy==1.5.0
prompt_toolkit==3.0.48
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
redis==5.1.1
s3transfer==0.10.3
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.14.0
